Refactoring the accounts tables
